% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert}
\alias{assert}
\title{Assert truthfulness of conditions before evaluation}
\usage{
assert(..., error_message = NULL)
}
\arguments{
\item{...}{expressions >= 1. If named the names are used
as error messages, otherwise R's internal error-messages are thrown}

\item{error_message}{character. An error message, supports \link{cli}-formatting.}
}
\value{
\link{NULL} if all statements in ... are \link{TRUE}
}
\description{
This function is a wrapper of \code{\link[=stopifnot]{stopifnot()}}, \code{\link[=tryCatch]{tryCatch()}} and
\code{\link[cli:cli_abort]{cli::cli_abort()}} and asserts the truthfulness of the passed expression(s).
}
\seealso{
\code{\link[=stopifnot]{stopifnot()}}, \code{\link[cli:cli_abort]{cli::cli_abort()}}, \code{\link[=tryCatch]{tryCatch()}}
}
\keyword{internal}
