% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{calendar}
\alias{calendar-package}
\alias{calendar}
\title{ics files with R}
\description{
\href{https://icalendar.org/}{iCalendar} is an open standard for
"exchanging calendar and scheduling information between users and computers".
Files adhering to this standard are save as \code{.ics} files.
}
\details{
The \code{ical} package is for interacting with such files
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atfutures/calendar}
  \item \url{https://atfutures.github.io/calendar/}
  \item \url{https://github.com/ATFutures/calendar}
  \item Report bugs at \url{https://github.com/ATFutures/calendar/issues}
}

}
\author{
\strong{Maintainer}: Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})

Authors:
\itemize{
  \item Layik Hama \email{layik.hama@gmail.com} (\href{https://orcid.org/0000-0003-1912-4890}{ORCID})
}

Other contributors:
\itemize{
  \item Ollie Lloyd \email{o.lloyd@doctors.org.uk} (\href{https://orcid.org/0000-0002-9385-1634}{ORCID}) [contributor]
}

}
