% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ic_attributes_vec}
\alias{ic_attributes_vec}
\title{Extract attributes from ical text}
\usage{
ic_attributes_vec(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
Extract attributes from ical text
}
\examples{
ic_attributes_vec(ical_example)
}
