% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_dataframe.R
\name{ic_bind_list}
\alias{ic_bind_list}
\title{Bind list of named vectors of variable length into data frame}
\usage{
ic_bind_list(x)
}
\arguments{
\item{x}{list of named vectors}
}
\description{
Based on: https://stackoverflow.com/questions/17308551/
}
\examples{
ic_bind_list(list(ic_vector(ical_example)))
ics_file <- system.file("extdata", "england-and-wales.ics", package = "ical")
ics_raw = readLines(ics_file)
x <- lapply(ic_list(ics_raw), function(x) {
  ic_vector(x)
})
ic_df <- ic_bind_list(x)
head(ic_df)
x <- lapply(ical_outlook, function(x) {
  ic_vector(x)
})
ic_bind_list(x)
}
