% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ic_char_event}
\alias{ic_char_event}
\title{Convert ical object to character strings of events}
\usage{
ic_char_event(ic, zulu = FALSE)
}
\arguments{
\item{ic}{object of class \code{ical}}

\item{zulu}{is the datetime in Zulu time?
\code{FALSE} by default, which means the calendar's current timezone
is used.}
}
\description{
Convert ical object to character strings of events
}
\examples{
ic <- ical(ical_example)
ic_char_event(ic)
ic_char_event(ic[c(1, 1), ])
}
