% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ic_character}
\alias{ic_character}
\title{Convert ical object to character strings of attributes}
\usage{
ic_character(ic)
}
\arguments{
\item{ic}{object of class \code{ical}}
}
\description{
Convert ical object to character strings of attributes
}
\examples{
ic <- ical(ical_example)
ic_character(ic)
identical(ical_example, ic_character(ic))
# ic_character(ic[c(1, 1), ]) # multiple events
}
