% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_date.R
\name{ic_date}
\alias{ic_date}
\title{Convert ical date into R date}
\usage{
ic_date(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
Convert ical date into R date
}
\examples{
ic_date("20120103")
}
