% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_event.R
\name{ic_event}
\alias{ic_event}
\title{Create ical object from properties_core inputs}
\usage{
ic_event(start = Sys.time(), end = Sys.time() + 1 * 60 * 60,
  summary = "R ical event")
}
\arguments{
\item{start}{start time}

\item{end}{end time}

\item{summary}{short outline of the event}
}
\value{
object of class ics
}
\description{
Create ical object from properties_core inputs
}
\examples{
ic_event()
}
