% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_extract.R
\name{ic_extract}
\alias{ic_extract}
\title{Extract contents of iCal fields}
\usage{
ic_extract(x, pattern)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}

\item{pattern}{A text string to search from (an ical field)}
}
\description{
Return formatted data from iCal fields
}
\examples{
ic_extract(ic_example, "DTSTART")
}
