% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_extract.R
\name{ic_extract_raw}
\alias{ic_extract_raw}
\title{Extract raw contents of iCal fields}
\usage{
ic_extract_raw(x, pattern)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}

\item{pattern}{A text string to search from (an ical field)}
}
\description{
Return info from iCal files, return raw unprocessed text.
}
\examples{
pattern = "TSTAMP"
ic_extract(ic_example, pattern)
}
