% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_event.R
\name{ic_guid}
\alias{ic_guid}
\title{Get an ical GUID}
\usage{
ic_guid()
}
\description{
Provided without any testing. Slight improvement from
\href{https://stackoverflow.com/a/10493590/2332101}{SO question}
}
\examples{
ic_guid()
}
