% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_read.R
\name{ic_read}
\alias{ic_read}
\title{Read ics file}
\usage{
ic_read(file)
}
\arguments{
\item{file}{ics file to read}
}
\value{
object of class ics
}
\description{
Read ics file
}
\examples{
f <- system.file("extdata", "england-and-wales.ics", package = "ical")
ics_df = ic_read(f)
head(ics_df)
}
