% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_readpy.R
\name{ic_readpy}
\alias{ic_readpy}
\title{Read ics file}
\usage{
ic_readpy(file)
}
\arguments{
\item{file}{ics file to read}
}
\value{
object of class ics
}
\description{
Read ics file
}
\examples{
f <- system.file("extdata", "england-and-wales.ics", package = "ical")
ics_list = ic_readpy(f)
ics_list
}
