% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_spec.R
\name{ic_spec}
\alias{ic_spec}
\title{View or download the ical specification}
\usage{
ic_spec(mode = "view", spec_url = "https://tools.ietf.org/rfc/rfc5545.txt")
}
\arguments{
\item{mode}{character string specifying whether to look at the spec (\code{mode = "view"}, the default)
or read it in (\code{mode = "read"})}

\item{spec_url}{the location of the latest version of the spec, from
tools.ietf.org/rfc/rfc5545.txt by default}
}
\description{
This function shows the spec underlying iCal files.
}
