% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_vector.R
\name{ic_vector}
\alias{ic_vector}
\title{Return a named vector from raw iCal text}
\usage{
ic_vector(x, pattern = paste0(ical::properties, collapse = "|"))
}
\arguments{
\item{x}{Lines read-in in from an iCal file}

\item{pattern}{A text string to search from (an ical field)}
}
\description{
This is designed to be a helper function for creating data frames
ical lists.
}
\examples{
x = ical_example[18:19]
ic_vector(x)
}
