% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{properties_ical}
\alias{properties_ical}
\title{ical default VCALENDAR properties in one line vectors.}
\format{
An object of class \code{character} of length 5.
}
\usage{
properties_ical
}
\description{
ical default VCALENDAR properties in one line vectors.
}
\examples{
# properties_ical <- ical_example[1:5]
# properties_ical
# properties_ical[2] <- "PRODID:-//ATFutures/ical //EN"
# usethis::use_data(properties_ical)
}
\keyword{datasets}
