% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{calendar}
\alias{calendar}
\title{ics files with R}
\description{
\href{https://icalendar.org/}{iCalendar} is an open standard for
"exchanging calendar and scheduling information between users and computers".
Files adhering to this standard are save as \code{.ics} files.
}
\details{
The \code{ical} package is for interacting with such files
}
