% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ic_attributes_vec}
\alias{ic_attributes_vec}
\title{Extract attributes from ical text}
\usage{
ic_attributes_vec(
  x = NULL,
  ic_attributes = c(BEGIN = "VCALENDAR", PRODID = "ATFutures/calendar", VERSION =
    "2.0", CALSCALE = "GREGORIAN", METHOD = "PUBLISH")
)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}

\item{ic_attributes}{Calendar attributes, e.g. as provided by \code{ic_attributes_vec()}.}
}
\description{
Extract attributes from ical text
}
\examples{
ic_attributes_vec() # default attributes (can be changed)
ic_attributes_vec(ical_example)
}
