% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_extract.R
\name{ic_char_datetime}
\alias{ic_char_datetime}
\title{Convert datetime object to character string}
\usage{
ic_char_datetime(x)
}
\arguments{
\item{x}{datetime object}
}
\description{
Convert datetime object to character string
}
\examples{
x <- structure(1533826800, class = c("POSIXct", "POSIXt"), tzone = "")
ic_char_datetime(x) == "20180809T160000Z"
}
