% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ic_character}
\alias{ic_character}
\title{Convert ical object to character strings of attributes}
\usage{
ic_character(ic, zulu = FALSE)
}
\arguments{
\item{ic}{object of class \code{ical}}

\item{zulu}{is the datetime in Zulu time?
\code{FALSE} by default, which means the calendar's current timezone
is used.}
}
\description{
Convert ical object to character strings of attributes
}
\examples{
ic <- ical(ical_example)
ic_character(ic)
identical(ical_example, ic_character(ic))
}
