% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_extract.R
\name{ic_datetime}
\alias{ic_datetime}
\title{Convert ical datetime into R datetime}
\usage{
ic_datetime(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
Convert ical datetime into R datetime
}
\examples{
ic_datetime("20180809T160000Z")
}
