% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_extract.R
\name{ic_datetime}
\alias{ic_datetime}
\title{Convert ical datetime into R datetime
Z at the end of an ical stamp stands of Zulu time
https://en.wikipedia.org/wiki/Coordinated_Universal_Time#Time_zones
which is UTC = GMT https://greenwichmeantime.com/info/zulu/}
\usage{
ic_datetime(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
Convert ical datetime into R datetime
Z at the end of an ical stamp stands of Zulu time
https://en.wikipedia.org/wiki/Coordinated_Universal_Time#Time_zones
which is UTC = GMT https://greenwichmeantime.com/info/zulu/
}
\examples{
ic_datetime("20180809T160000Z")
}
