% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_event.R
\name{ic_event}
\alias{ic_event}
\title{Create ical object from properties_core inputs}
\usage{
ic_event(uid = ic_guid(),
  start_time = as.POSIXct(round.POSIXt(Sys.time(), units = "hours")),
  end_time = as.POSIXct(round.POSIXt(Sys.time(), units = "hours") + 1 *
  60 * 60), summary = "ical event", more_properties = FALSE,
  event_properties = stats::setNames(rep(NA,
  length(setdiff(ical::properties, ical::properties_core))), nm =
  setdiff(ical::properties, ical::properties_core)))
}
\arguments{
\item{uid}{the unique id of the event, by default generated by \code{ic_uid()}}

\item{start_time}{start time, by default the current time to the nearest hour}

\item{end_time}{end time, by defult and hour (3600 seconds) after \code{start_time}}

\item{summary}{short outline of the event}

\item{more_properties}{add placeholder columns for properties in addition to \code{properties_core},
dy default \code{FALSE}}

\item{event_properties}{named vector of additional properties to include, by default
with names \code{ical::properties} containing \code{NA}s to be subsequently populated}
}
\value{
object of class ics
}
\description{
Create ical object from properties_core inputs
}
\examples{
ic_event()
ic_event(more_properties = TRUE)
}
