% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_event.R
\name{ic_event}
\alias{ic_event}
\title{Create ical object from properties_core inputs}
\format{in case of start_time and end_time being character values, a format must be provided.}
\usage{
ic_event(uid = ic_guid(),
  start_time = as.POSIXct(round.POSIXt(Sys.time(), units = "hours")),
  end_time = as.POSIXct(round.POSIXt(Sys.time(), units = "hours") + 1L *
  60L * 60L), format = "\%Y-\%m-\%d \%H:\%M", summary = "ical event",
  more_properties = FALSE, event_properties = ical::properties)
}
\arguments{
\item{uid}{the unique id of the event, by default generated by \code{ic_uid()}}

\item{start_time}{start time, by default the current time to the nearest hour}

\item{end_time}{end time, by defult and hour (3600 seconds) after \code{start_time}}

\item{format}{required if start_time and end_time are vectors and are not of datetime format
"\%Y-\%m-\%d \%H:\%M", you can use \code{ical::formats} object for convenience.}

\item{summary}{short outline of the event}

\item{more_properties}{add placeholder columns for properties in addition to \code{properties_core},
dy default \code{FALSE}}

\item{event_properties}{named vector of additional properties to include, by default
with names \code{ical::properties} containing \code{NA}s to be subsequently populated}
}
\value{
object of class ics
}
\description{
Create an ical event using either POSIXct type or character type with format parameters.
}
\examples{
ic_event()
ic_event(more_properties = TRUE)
ic_event(start_time = "18-10-12", end_time = "18-10-13", format = ical::formats$`yy-mm-dd`)
}
