% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_event.R
\name{ic_event_raw}
\alias{ic_event_raw}
\title{Create ical object using character dates}
\usage{
ic_event_raw(start = "20180927 13:00", end = "20180927 15:00",
  summary = "R ical minihack", format = "\%Y\%m\%d \%H:\%M")
}
\arguments{
\item{start}{start as character}

\item{end}{end as character}

\item{summary}{short outline of the event}

\item{format}{strptime format}
}
\value{
object of class ics
}
\description{
Create ical object using character dates
}
\examples{
ic_event_raw()
}
