% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ic_example}
\alias{ic_example}
\title{Minimal example of raw ical file (already read-in from file)}
\format{An object of class \code{character} of length 22.}
\usage{
ic_example
}
\description{
Minimal example of raw ical file (already read-in from file)
}
\examples{
# download.file("long_url...", "inst/extdata/example.ics")
ic_example = readLines(system.file("extdata", "example.ics", package = "ical"))
# usethis::use_data(ic_example)
}
\keyword{datasets}
