% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_find.R
\name{ic_find}
\alias{ic_find}
\title{Find contents of iCal fields}
\usage{
ic_find(x, pattern)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}

\item{pattern}{A text string to search from (an ical field)}
}
\description{
Find contents of iCal fields
}
\examples{
pattern = "DTSTAMP"
ic_find(ic_example, pattern)
}
