% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ic_properties_core}
\alias{ic_properties_core}
\title{The key 'properties' that are allowed in ical files}
\format{An object of class \code{list} of length 4.}
\usage{
ic_properties_core
}
\description{
The key 'properties' that are allowed in ical files
}
\examples{
ic_properties_core = list(
"UID"="UID",
"DTSTART"="DTSTART",
"DTEND"="DTEND",
"SUMMARY"="SUMMARY"
)
}
\keyword{datasets}
