% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ical}
\alias{ical}
\title{Create object of class ical}
\usage{
ical(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
Create object of class ical
}
\examples{
ic <- ical(ical_example)
class(ic)
attributes(ic)
}
