% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R, R/package.R
\docType{package}
\name{ical}
\alias{ical}
\alias{ical-package}
\title{Create object of class ical}
\usage{
ical(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
\href{https://icalendar.org/}{iCalendar} is an open standard for
"exchanging calendar and scheduling information between users and computers".
Files adhering to this standard are save as \code{.ics} files.
}
\details{
The \code{ical} package is for interacting with such files
}
\examples{
ic <- ical(ical_example)
class(ic)
attributes(ic)
}
