% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical.R
\name{ical}
\alias{ical}
\title{Create object of class ical}
\usage{
ical(x)
}
\arguments{
\item{x}{Lines read-in in from an iCal file}
}
\description{
Create object of class ical
}
\examples{
# ical from .ics characters:
class(ical_example)
ic <- ical(ical_example)
attributes(ic)
class(ic)
# ical from data frame:
ic_df <- data.frame(ic)
ic2 <- ical(ic_df)
class(ic2)
}
